chore(): 1.0.13

+++ setup.sh
@@ -9,14 +9,17 @@
 #        "               ""
 #        bash package manager
 
+VERSION="1.0.13"
+TAG=${TAG:-$VERSION}
+BRANCH=${BRANCH:-$TAG}
 REMOTE=${REMOTE:-https://github.com/bpkg/bpkg.git}
 TMPDIR=${TMPDIR:-/tmp}
-DEST=${DEST:-$TMPDIR/bpkg-master}
+DEST=${DEST:-$TMPDIR/bpkg-$BRANCH}
 
 ## test if command exists
 ftest () {
-  echo "  info: Checking for ${1}..."
-  if ! type -f "${1}" > /dev/null 2>&1; then
+  echo "  info: Checking for $1..."
+  if ! type -f "$1" > /dev/null 2>&1; then
     return 1
   else
     return 0
@@ -26,8 +29,8 @@ ftest () {
 ## feature tests
 features () {
   for f in "${@}"; do
-    ftest "${f}" || {
-      echo >&2 "  error: Missing \`${f}'! Make sure it exists and try again."
+    ftest "$f" || {
+      echo >&2 "  error: Missing \`$f'! Make sure it exists and try again."
       return 1
     }
   done
@@ -43,13 +46,15 @@ setup () {
   ## build
   {
     echo
-    cd "${TMPDIR}" || exit
     echo "  info: Creating temporary files..."
-    test -d "${DEST}" && { echo "  warn: Already exists: '${DEST}'"; }
-    rm -rf "${DEST}"
-    echo "  info: Fetching latest 'bpkg'..."
-    git clone --depth=1 "${REMOTE}" "${DEST}" > /dev/null 2>&1
-    cd "${DEST}" || exit
+    cd "$TMPDIR" || exit
+    test -d "$DEST" && { echo "  warn: Already exists: '$DEST'"; }
+    rm -rf "$DEST"
+
+    echo "  info: Fetching 'bpkg@$BRANCH'..."
+    git clone --depth=1 --branch "$BRANCH" "$REMOTE" "$DEST" > /dev/null 2>&1
+    cd "$DEST" || exit
+
     echo "  info: Installing..."
     echo
     make_install
@@ -61,7 +66,7 @@ setup () {
 ## make targets
 BIN="bpkg"
 if [ -z "$PREFIX" ]; then
-  if [ "$USER" == "root" ]; then
+  if [ "$(whoami)" == "root" ]; then
     PREFIX="/usr/local"
   else
     PREFIX="$HOME/.local"
@@ -84,6 +89,7 @@ CMDS+=("suggest")
 CMDS+=("term")
 CMDS+=("update")
 CMDS+=("utils")
+CMDS+=("realpath")
 
 make_install () {
   local source
@@ -97,17 +103,19 @@ make_install () {
 
   if [ -f "$source" ]; then
     install "$source" "$PREFIX/bin/$BIN"
-    else
-      install "$BIN" "$PREFIX/bin"
+  else
+    install "$BIN" "$PREFIX/bin"
   fi
 
   for cmd in "${CMDS[@]}"; do
-    source=$(<"$BIN-$cmd")
+    if test -f "$BIN-$cmd"; then
+      source=$(<"$BIN-$cmd")
 
-    if [ -f "$source" ]; then
-      install "$source" "$PREFIX/bin/$BIN-$cmd"
-    else
-      install "$BIN-$cmd" "$PREFIX/bin"
+      if [ -f "$source" ]; then
+        install "$source" "$PREFIX/bin/$BIN-$cmd"
+      else
+        install "$BIN-$cmd" "$PREFIX/bin"
+      fi
     fi
 
   done
@@ -119,8 +127,10 @@ make_uninstall () {
   echo "    rm: $PREFIX/bin/$BIN'"
   rm -f "$PREFIX/bin/$BIN"
   for cmd in "${CMDS[@]}"; do
-    echo "    rm: $PREFIX/bin/$BIN-$cmd'"
-    rm -f "$PREFIX/bin/$BIN-$cmd"
+    if test -f "$PREFIX/bin/$BIN-$cmd"; then
+      echo "    rm: $PREFIX/bin/$BIN-$cmd'"
+      rm -f "$PREFIX/bin/$BIN-$cmd"
+    fi
   done
   return $?
 }
@@ -131,8 +141,10 @@ make_link () {
   echo "  link: '$PWD/$BIN' -> '$PREFIX/bin/$BIN'"
   ln -s "$PWD/$BIN" "$PREFIX/bin/$BIN"
   for cmd in "${CMDS[@]}"; do
-    echo "  link: '$PWD/$BIN-$cmd' -> '$PREFIX/bin/$BIN-$cmd'"
-    ln -s "$PWD/$BIN-$cmd" "$PREFIX/bin/$BIN-$cmd"
+    if test -f "$PWD/$BIN-$cmd"; then
+      echo "  link: '$PWD/$BIN-$cmd' -> '$PREFIX/bin/$BIN-$cmd'"
+      ln -s "$PWD/$BIN-$cmd" "$PREFIX/bin/$BIN-$cmd"
+    fi
   done
   return $?
 }
